app/requirements.txt
cryptography==44.0.0
fastapi==0.115.8
uvicorn[standard]==0.34.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
